Change ClassHierarchy to be a class instead of an interface
